refactor(api): перенос констант ролей в общий пакет types
