+++ apps/api/src/common/constants/roles.constants.ts
@@ -1,9 +1,4 @@
-export const TEAM_ROLES = {
-	OWNER: 'OWNER',
-	ADMIN: 'ADMIN',
-	MEMBER: 'MEMBER',
-} as const
-
-export type TeamRole = (typeof TEAM_ROLES)[keyof typeof TEAM_ROLES]
+export { TEAM_ROLES } from '@repo/types'
+export type { TeamRole } from '@repo/types'
 
 export const ROLES_KEY = 'roles'
