+++ packages/types/src/entry.ts
@@ -11,3 +11,5 @@ export { loginRequestSchema } from './auth/schema/login-request.schema'
 export type { LoginRequest } from './auth/schema/login-request.schema'
 export { authResponseSchema } from './auth/schema/auth-response.schema'
 export type { AuthResponse } from './auth/schema/auth-response.schema'
+export { TEAM_ROLES } from './teams/team-role.constants'
+export type { TeamRole } from './teams/team-role.constants'
