+++ packages/types/src/teams/team-role.constants.ts
@@ -1,0 +1,7 @@
+export const TEAM_ROLES = {
+	OWNER: 'OWNER',
+	ADMIN: 'ADMIN',
+	MEMBER: 'MEMBER',
+} as const
+
+export type TeamRole = (typeof TEAM_ROLES)[keyof typeof TEAM_ROLES]
